requirements.txt
APScheduler==3.6.3
backports.zoneinfo==0.2.1
cachetools==4.2.2
certifi==2022.6.15
python-telegram-bot==13.12
pytube==12.1.0
pytube3==9.6.4
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
six==1.16.0
tornado==6.1
typing_extensions==4.2.0
tzdata==2022.1
tzlocal==4.2
